Exceptios middleware implemented, some errors with cart were fixed

+++ OrderService/OrderService.Application/DTO/CartDTO.cs
@@ -5,6 +5,6 @@ namespace OrderService.Application.DTO;
 public class CartDTO
 {
 	public int Count { get; set; }
-	public int TotalCost { get; set; }
-	public List<Product> Products {  get; set; } 
+	public double TotalCost { get; set; }
+	public List<Product> Products { get; set; } 
 }
